Optimize the case when just sending fin without no stream data
Previously, we send 0 length DATA frame if we need to send just fin.
In HTTP/3, 0 length DATA is just a waste of bytes.  This change
removes the need for it and enables app to send fin without extra
bytes.

Unfortunately, this requires that app has to call
nghttp3_conn_add_write_offset as long as nghttp3_conn_writev_stream
fills stream_id with other than -1 even if its return value is 0.