Fix overflow of min_dtable_update
size_t is 32-bit on 32-bit machines and min_dtable_update was assigned NGHTTP3_QPACK_INT_MAX, which is 64 bits. This failed the tests on 32-bit machines.

There are a few other places in nghttp3_qpack where uint64_t values are assigned to size_t variables and need to be fixed.