Handle decoder buffer overflow
It is possible for a remote endpoint not to give any stream flow
control credits to QPACK decoder stream and keep resetting streams.
If this happens, QPACK decoder stream piles up because decoder has to
write stream cancellation and eventually memory is exhausted.  To
handle this situation, set a limit to the maximum size of decoder
stream buffer which is proportional to the number of the concurrent
streams that the remote endpoint can open.  If the limit is exceeded,
drop the connection.