Remove server push
Remove server push because it is not used widely and failed to show
any measurable improvements in HTTP/2 deployment.  Yet, it is very
much complex and requires lots of code and careful management of
PUSH_PROMISE and pushed stream.  Since one of the major browser vendor
announced that it does not support HTTP/3 server push, its usage will
be further decreased.  It is better to remove it now and hope that the
next major HTTP removes server push entirely.