Add fin parameter to nghttp3_end_headers callback
Add fin parameter to nghttp3_end_headers callback to tell an
application that the stream has ended with this header block.
Previously, the application cannot know the stream has ended or not
and has to wait for nghttp3_end_stream callback to be invoked or
proceed as if the message might have a body.  This is not really
useful because if one chooses latter, you might need
transfer-encoding: chunked when forwarding the request in HTTP/1.1 hop
even if the method is GET.