Ignore priority header field in trailer fields
RFC9110 Section 16.3.2 states that each header field specification
needs to document that whether or not the header field is allowed in
trailers, and by default it will not be allowed.  RFC9218 does not say
anything about trailers, which means that priority header field is not
allowed in trailers.  Therefore, we ignore priority header field.