qpack: Add indexing strategy
The default indexing strategy for fields that are not defined in
nghttp3_qpack_token is not index those fields.  The common fields are
often included in nghttp3_qpack_token, so this keeps the dynamic table
from discarding those fields by inserting the unpopular fields.
NGHTTP3_NV_FLAG_TRY_INDEX allows application to selectively index
those fields.  This is a good strategy for the clients and the origin
servers that have a best knowledge of the HTTP fields they transmit.
However, for intermediaries, they might not have such knowledge, and
it is often better to index everything.  They can achieve this with
newly introduced NGHTTP3_QPACK_INDEXING_STRAT_EAGER, which tells QPACK
encoder to encode all fields not defined in nghttp3_qpack_token.