Add nghttp3_recv_settings2 and deprecate nghttp3_recv_settings
When nghttp3_recv_settings was introduced, nghttp3_settings just
contained values that could be sent or received in HTTP/3 SETTINGS
frame except for a single field.  The choice of the struct was just a
convenience.

Since then we have added more fields to nghttp3_settings that are not
send or received in SETTINGS frame because nghttp3_settings is the
configuration for nghttp3_conn object, not just for HTTP/3 SETTINGS
values.  Now it is apparent that nghttp3_settings is not appropriate
for use when notifying the contents of received SETTINGS frame.

To fix this issue, this commit adds new struct named
nghttp3_proto_settings which only contains HTTP/3 SETTINGS values that
this library can recognize.  New callback function
nghttp3_recv_settings2 is also added to notify the incoming SETTINGS
to the application with this new struct.

nghttp3_recv_setting is now deprecated in favor of
nghttp3_recv_settings2.  The new HTTP/3 settings that might be added
in the future will not be notified with this deprecated callback.